Improve error prone implementation of CTCSS to TG map
Was using float as index into map which can lead to problems when
rounding errors are present. Now using uint16_t instead.

+++ src/svxlink/svxlink/Logic.h
@@ -282,7 +282,7 @@ class Logic : public LogicBase
     DtmfDigitHandler                *dtmf_digit_handler;
     Async::Pty                      *state_pty;
     Async::Pty                      *dtmf_ctrl_pty;
-    std::map<float, uint32_t>       m_ctcss_to_tg;
+    std::map<uint16_t, uint32_t>    m_ctcss_to_tg;
 
     void loadModules(void);
     void loadModule(const std::string& module_name);
